build: Fix remove extract dir if non-existent

+++ xtask/src/nuke/sources.rs
@@ -405,7 +405,9 @@ async fn install_windows(major: usize, installer: &Path, install_path: &Path) ->
         }
         #[cfg(target_os = "windows")]
         {
-            tokio::fs::remove_dir_all(installer.parent().unwrap().join("extract")).await?;
+            if installer.parent().unwrap().join("extract").is_dir() {
+                tokio::fs::remove_dir_all(installer.parent().unwrap().join("extract")).await?;
+            };
             let install_directory = installer
                 .parent()
                 .unwrap()
